pyproject.toml
[project]
name = "hexlet-price-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "django-allauth>=65.9.0",
    "django-bootstrap5>=25.1",
    "django-widget-tweaks>=1.5.0",
    "dotenv>=0.9.9",
    "gunicorn>=23.0.0",
    "inertia-django>=1.2.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
    "ruff>=0.12.2",
    "telethon>=1.40.0",
]
